fix: move the applyScriptCondition func to builder-shared
