+++ packages/builder/builder-shared/src/apply/index.ts
@@ -3,3 +3,4 @@ export * from './output';
 export * from './resolve';
 export * from './basic';
 export * from './hmr';
+export * from './babel';
