+++ packages/builder/builder-webpack-provider/src/plugins/babel.ts
@@ -10,7 +10,7 @@ import {
   createVirtualModule,
   getBrowserslistWithDefault,
   getSharedPkgCompiledPath,
-  BuilderContext,
+  applyScriptCondition,
 } from '@modern-js/builder-shared';
 
 import type { WebpackChain, BuilderPlugin, NormalizedConfig } from '../types';
@@ -28,33 +28,6 @@ export const getUseBuiltIns = (config: NormalizedConfig) => {
   }
   return polyfill;
 };
-
-export function applyScriptCondition({
-  rule,
-  config,
-  context,
-  includes,
-  excludes,
-}: {
-  rule: WebpackChain.Rule;
-  config: NormalizedConfig;
-  context: BuilderContext;
-  includes: (string | RegExp)[];
-  excludes: (string | RegExp)[];
-}) {
-  // compile all folders in app directory, exclude node_modules
-  rule.include.add({
-    and: [context.rootPath, { not: /node_modules/ }],
-  });
-
-  [...includes, ...(config.source.include || [])].forEach(condition => {
-    rule.include.add(condition);
-  });
-
-  [...excludes, ...(config.source.exclude || [])].forEach(condition => {
-    rule.exclude.add(condition);
-  });
-}
 
 export const builderPluginBabel = (): BuilderPlugin => ({
   name: 'builder-plugin-babel',
